fix(cli): remove standalone trade preset commands
