+++ cmd/smoke-test/main.go
@@ -314,8 +314,6 @@ func buildStaticCases(date string) map[string]smokeCase {
 		"trade level-touches":             readOnly("trade level-touches", "AAPL", "--days", "14", "--length", "10"),
 		"trade levels":                    readOnly("trade levels", "AAPL", "--trade-level-count", "5"),
 		"trade list":                      readOnly("trade list", "AAPL", "--days", "1"),
-		"trade preset-tickers":            readOnly("trade preset-tickers", "--preset", "All Trades"),
-		"trade presets":                   readOnly("trade presets"),
 		"trade sentiment":                 readOnly("trade sentiment", "--days", "5"),
 		"volume ah-institutional":         readOnly("volume ah-institutional", "--date", date, "--length", "10"),
 		"volume institutional":            readOnly("volume institutional", "--date", date, "--length", "10"),
@@ -425,8 +423,6 @@ func orderedCommands(commands map[string]struct{}) []string {
 		"report top-30-rank-10x-99th",
 		"report phantom-trades",
 		"report offsetting-trades",
-		"trade presets",
-		"trade preset-tickers",
 		"market snapshots",
 		"market exhaustion",
 		"market earnings",
